+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/VRCQuestToolsAvatarProcessor.cs
@@ -5,7 +5,6 @@
 using KRT.VRCQuestTools.Components;
 using KRT.VRCQuestTools.Models;
 using KRT.VRCQuestTools.Utils;
-using System.Linq;
 using System.Text.RegularExpressions;
 using UnityEditor;
 using UnityEngine;
@@ -48,14 +47,13 @@ namespace KRT.VRCQuestTools
                 r.RemoveVertexColor();
             }
 
+#if !VQT_HAS_NDMF
             var components = avatarGameObject.GetComponentsInChildren<VRCQuestToolsEditorOnly>(true);
-#if VQT_HAS_NDMF
-            components = components.Where(c => !(c is IVRCQuestToolsNdmfComponent)).ToArray();
-#endif
             foreach (var c in components)
             {
                 Object.DestroyImmediate(c);
             }
+#endif
 
             return true;
         }