+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add `VQT Platform GameObject Remover` component to remove game objects depending on the build platform (NDMF is required).
+- Add warning when using NDMF-required components but NDMF is not installed.
 
 ### Fixed
 - Fixed an issue where the converter fails when the avatar object name starts or ends with whitespaces.